requirements.txt
Django==2.2
django-extensions==2.1.6
django-taggit==1.1.0
django-test-plus==1.1.1
djangorestframework==3.9.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
